split schema inspection, tweak paths
